fix(ruler): 修改类型定义
