--- packages/core/ruler/type.d.ts
@@ -1,50 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import type CanvasRuler, { Rect } from './ruler';
-
-declare module 'fabric/fabric-impl' {
-  type EventNameExt = 'removed' | EventName;
-
-  export interface Canvas {
-    _setupCurrentTransform(e: Event, target: fabric.Object, alreadySelected: boolean): void;
-  }
-
-  export interface IObservable<T> {
-    on(
-      eventName: 'guideline:moving' | 'guideline:mouseup',
-      handler: (event: { e: Event; target: fabric.GuideLine }) => void
-    ): T;
-    on(events: { [key: EventName]: (event: { e: Event; target: fabric.GuideLine }) => void }): T;
-  }
-
-  export interface IGuideLineOptions extends ILineOptions {
-    axis: 'horizontal' | 'vertical';
-  }
-
-  export interface IGuideLineClassOptions extends IGuideLineOptions {
-    canvas: {
-      setActiveObject(object: fabric.Object | fabric.GuideLine, e?: Event): Canvas;
-      remove<T>(...object: (fabric.Object | fabric.GuideLine)[]): T;
-    } & Canvas;
-    activeOn: 'down' | 'up';
-    initialize(xy: number, objObjects: IGuideLineOptions): void;
-    callSuper(methodName: string, ...args: unknown[]): any;
-    getBoundingRect(absolute?: boolean, calculate?: boolean): Rect;
-    on(eventName: EventNameExt, handler: (e: IEvent<MouseEvent>) => void): void;
-    off(eventName: EventNameExt, handler?: (e: IEvent<MouseEvent>) => void): void;
-    fire<T>(eventName: EventNameExt, options?: any): T;
-    isPointOnRuler(e: MouseEvent): 'horizontal' | 'vertical' | false;
-    bringToFront(): fabric.Object;
-    isHorizontal(): boolean;
-  }
-
-  export interface GuideLine extends Line, IGuideLineClassOptions {}
-
-  export class GuideLine extends Line {
-    constructor(xy: number, objObjects?: IGuideLineOptions);
-    static fromObject(object: any, callback: any): void;
-  }
-
-  export interface StaticCanvas {
-    ruler: InstanceType<typeof CanvasRuler>;
-  }
-}
